# src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Swagger Demo Project",
    "description": "Implementation of Swagger with TypeScript"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "API routes 2"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/spam": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "example": "any"
                  },
                  "countryCode": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "description": "To create user",
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "user token"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "user token"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "To create a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/createUser"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#components/schemas/createUser"
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "If authenticated, will return token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "token"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "token"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "Endpoint to login and to get the token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/login"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#components/schemas/login"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "examples": {
      "login": {
        "value": {
          "email": "atulvinod1911@gmail.com",
          "password": "mypassword"
        }
      }
    },
    "schemas": {
      "login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "email"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "login"
        }
      },
      "createUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "email"
          },
          "name": {
            "type": "string",
            "example": "name"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "phoneNumber": {
            "type": "string",
            "example": "phoneNumber"
          },
          "countryCode": {
            "type": "string",
            "example": "countryCode"
          }
        },
        "required": [
          "email",
          "name",
          "password",
          "phoneNumber",
          "countryCode"
        ],
        "xml": {
          "name": "createUser"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}